testing: add testcase for besu quirk, improve minimizer
