+++ evms/besu.go
@@ -17,17 +17,11 @@
 package evms
 
 import (
-	"bufio"
-	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
-	"os"
 	"os/exec"
 	"strings"
-
-	"github.com/ethereum/go-ethereum/eth/tracers/logger"
 )
 
 // BesuVM is s Evm-interface wrapper around the `evmtool` binary, based on Besu.
@@ -59,9 +53,6 @@ func (evm *BesuVM) RunStateTest(path string, out io.Writer, speedTest bool) (str
 	} else {
 		cmd = exec.Command(evm.path, "--nomemory", "--json",
 			"state-test", path) // exclude memory
-		// For running this via docker, this is the 'raw' base command
-		//docker run --rm  -i -v ~/yolov2:/yolov2/ hyperledger/besu-evmtool:develop --Xberlin-enabled true state-test  /yolov2/tests/$f
-
 	}
 	if stdout, err = cmd.StdoutPipe(); err != nil {
 		return cmd.String(), err
@@ -81,10 +72,8 @@ func (vm *BesuVM) Close() {
 
 func (vm *BesuVM) GetStateRoot(path string) (root, command string, err error) {
 	// Run without tracing
-	cmd := exec.Command(vm.path, "--nomemory",
-		"state-test", path)
+	cmd := exec.Command(vm.path, "--nomemory", "state-test", path)
 
-	/// {"output":"","gasUsed":"0x798765","time":342028058,"test":"00000000-storagefuzz-0","fork":"Berlin","d":0,"g":0,"v":0,"postHash":"0xbaeec41171e943575740bb99cdd8ffd45fa6207bf170f99c1a61425069ffae97","postLogsHash":"0x1dcc4de8dec75d7aab85b567b6ccd41ad312451b948a7413f0a142fd40d49347","pass":false}
 	data, err := cmd.CombinedOutput()
 	if err != nil {
 		return "", cmd.String(), err
@@ -98,65 +87,8 @@ func (vm *BesuVM) GetStateRoot(path string) (root, command string, err error) {
 	return "", cmd.String(), errors.New("no stateroot found")
 }
 
-type besuStateRoot struct {
-	StateRoot string `json:"postHash"`
-}
-
 // feed reads from the reader, does some geth-specific filtering and
 // outputs items onto the channel
 func (evm *BesuVM) Copy(out io.Writer, input io.Reader) {
-	var stateRoot stateRoot
-	scanner := bufio.NewScanner(input)
-	// We use a larger scanner buffer for besu: it does not have a way to
-	// disable 'returndata', which can become larger than fits into a default
-	// scanner buffer
-	buf := make([]byte, 16*1024*1024)
-	scanner.Buffer(buf, cap(buf))
-	for scanner.Scan() {
-		data := scanner.Bytes()
-		var elem logger.StructLog
-		// Besu (like Nethermind) sometimes report a negative refund
-		if i := bytes.Index(data, []byte(`"refund":-`)); i > 0 {
-			// we can just make it positive, it will be zeroed later
-			data[i+9] = byte(' ')
-		}
-
-		err := json.Unmarshal(data, &elem)
-		if err != nil {
-			fmt.Printf("besu err: %v, line\n\t%v\n", err, string(data))
-			continue
-		}
-		// If the output cannot be marshalled, all fields will be blanks.
-		// We can detect that through 'depth', which should never be less than 1
-		// for any actual opcode
-		if elem.Depth == 0 {
-			if stateRoot.StateRoot == "" {
-				var tempRoot besuStateRoot
-				if err := json.Unmarshal(data, &tempRoot); err == nil {
-					// Besu calls state root "rootHash"
-					stateRoot.StateRoot = tempRoot.StateRoot
-				}
-			}
-			//fmt.Printf("%v\n", string(data))
-			// For now, just ignore these
-			continue
-		}
-		// When geth encounters end of code, it continues anyway, on a 'virtual' STOP.
-		// In order to handle that, we need to drop all STOP opcodes.
-		if elem.Op == 0x0 {
-			continue
-		}
-		RemoveUnsupportedElems(&elem)
-
-		jsondata, _ := json.Marshal(elem)
-		if _, err := out.Write(append(jsondata, '\n')); err != nil {
-			fmt.Fprintf(os.Stderr, "Error writing to out: %v\n", err)
-			return
-		}
-	}
-	root, _ := json.Marshal(stateRoot)
-	if _, err := out.Write(append(root, '\n')); err != nil {
-		fmt.Fprintf(os.Stderr, "Error writing to out: %v\n", err)
-		return
-	}
+	besuCopyUntilEnd(out, input)
 }
