+++ evms/testdata/cases/00000006-naivefuzz-0.json
