+++ evms/testdata/traces/00000006-naivefuzz-0.json.besu.stdout.txt
