+++ evms/testdata/traces/00000006-naivefuzz-0.json.erigon.stderr.txt
