+++ evms/testdata/traces/00016209-naivefuzz-0.json.geth.stdout.txt
