+++ evms/testdata/traces/statetest1.json.erigon.stdout.txt
