+++ evms/testdata/traces/statetest1.json.geth.stderr.txt
