+++ evms/testdata/traces/statetest1.json.geth.stdout.txt
